Refactor where foreign key check functionality lives
